test(platform-ui): move L2 integration tests to src/test/integration/
Reflects 5-level testing pyramid from AGENTS.md. Files in src/test/
mixed integration tests (route handlers with mocked services) with
configuration (setup.ts, mocks/). Pulling integration tests into a
dedicated subdirectory makes the level explicit.

- src/test/*.test.ts -> src/test/integration/
- mcp-journey.test.ts -> mcp-integration.test.ts (the "journey" naming
  was misleading — these don't run against real services)
- mcp-oauth-journey.test.ts -> mcp-oauth-integration.test.ts
- Fix relative imports + __dirname-based path resolution after the move
- Update 2 unrelated comment refs that pointed at old paths

src/test/setup.ts + mocks/ stay (config, not tests).
Co-located unit tests in src/**/__tests__/ unchanged.

658 tests pass.
