+++ packages/platform-ui/src/test/integration/api-plugins-route.test.ts
@@ -2,7 +2,7 @@ import { describe, it, expect, vi, beforeEach } from 'vitest';
 import type { PluginCapabilityMetadata } from '@mediforce/platform-core';
 
 // Mock getPlatformServices before importing the route
-vi.mock('../lib/platform-services.js', () => {
+vi.mock('../../lib/platform-services.js', () => {
   const mockPlugins = new Map<string, { metadata?: PluginCapabilityMetadata }>();
   return {
     getPlatformServices: () => ({
@@ -22,13 +22,13 @@ describe('GET /api/plugins', () => {
   let mockPlugins: Map<string, { metadata?: PluginCapabilityMetadata }>;
 
   beforeEach(async () => {
-    const mod = await import('../lib/platform-services.js');
+    const mod = await import('../../lib/platform-services.js');
     mockPlugins = (mod as unknown as { _mockPlugins: Map<string, { metadata?: PluginCapabilityMetadata }> })._mockPlugins;
     mockPlugins.clear();
   });
 
   it('[DATA] returns 200 with plugins array', async () => {
-    const { GET } = await import('../app/api/plugins/route.js');
+    const { GET } = await import('../../app/api/plugins/route.js');
     const response = await GET();
     expect(response.status).toBe(200);
     const body = await response.json();
@@ -46,7 +46,7 @@ describe('GET /api/plugins', () => {
     };
     mockPlugins.set('sc/compliance-analyzer', { metadata: testMetadata });
 
-    const { GET } = await import('../app/api/plugins/route.js');
+    const { GET } = await import('../../app/api/plugins/route.js');
     const response = await GET();
     const body = await response.json();
 
@@ -59,7 +59,7 @@ describe('GET /api/plugins', () => {
   it('[DATA] returns plugin without metadata as undefined', async () => {
     mockPlugins.set('basic-plugin', {});
 
-    const { GET } = await import('../app/api/plugins/route.js');
+    const { GET } = await import('../../app/api/plugins/route.js');
     const response = await GET();
     const body = await response.json();
