+++ packages/platform-ui/src/test/integration/mcp-integration.test.ts
