+++ packages/platform-ui/src/test/integration/mcp-oauth-integration.test.ts
