+++ packages/platform-ui/src/test/integration/smoke.test.ts
